LoopTools.jl

For more information about the loop integrals, see the LoopTools manual.

The LoopTools Visitor Center

LoopTools.A0

LoopTools.A0i

LoopTools.Aget

LoopTools.B0

LoopTools.B00

LoopTools.B001

LoopTools.B0i

LoopTools.B11

LoopTools.B111

LoopTools.Bget

LoopTools.C0

LoopTools.C0i

LoopTools.Cget

LoopTools.D0

LoopTools.D0i

LoopTools.DB0

LoopTools.DB00

LoopTools.DB1

LoopTools.DB11

LoopTools.DRResult

LoopTools.Dget

LoopTools.E0

LoopTools.E0i

LoopTools.Eget

LoopTools.Li2

LoopTools.Li2omx

LoopTools.aget

LoopTools.agetsym

LoopTools.aput!

LoopTools.bget

LoopTools.bgetsym

LoopTools.bput!

LoopTools.cget

LoopTools.cgetsym

LoopTools.clearcache

LoopTools.cput!

LoopTools.dget

LoopTools.dgetsym

LoopTools.dput!

LoopTools.eget

LoopTools.egetsym

LoopTools.eput!

LoopTools.getcmpbits

LoopTools.getdebugkey

LoopTools.getdelta

LoopTools.getdiffeps

LoopTools.getlambda

LoopTools.getmaxdev

LoopTools.getminmass

LoopTools.getmudim

LoopTools.getuvdiv

LoopTools.getversionkey

LoopTools.getzeroeps

LoopTools.ltexi

LoopTools.markcache

LoopTools.restorecache

LoopTools.setcmpbits

LoopTools.setdebugkey

LoopTools.setdebugrange

LoopTools.setdelta

LoopTools.setdiffeps

LoopTools.setlambda

LoopTools.setmaxdev

LoopTools.setminmass

LoopTools.setmudim

LoopTools.setuvdiv

LoopTools.setversionkey

LoopTools.setzeroeps

LoopTools.A0 — Method

A0(m^2)

the scalar one-point one-loop function

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int d^D q \frac{1}{q^2-m^2}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

LoopTools.A0i — Function

A0i(id, m^2)

one-point tensor coefficient for id

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int d^D q \frac{\{1, g_{μν} \} }{q^2-m^2}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$`.

Special cases:

id	Int	Description

aa0	1	scalar one-point one-loop function, i.e., A0(m^2)

aa00	4	coefficient of $g_{μν}$

LoopTools.Aget — Function

Aget(m^2; val_only = false)

the finite piece of all one-point tensor coefficients. See also aget.

val_only = false, return a NamedTuple; otherwise return an NTuple.

LoopTools.B0 — Function

B0(p^2, m1^2, m2^2)

the scalar two-point one-loop function

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{d^D q }{(q^2-m_1^2)\left[(q+p)^2-m_2^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

LoopTools.B00 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.B001 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.B0i — Method

B0i(id, p^2, m1^2, m2^2)

two-point tensor coefficient for id

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{({\rm numerator})\, d^D q }{(q^2-m_1^2)\left[(q+p)^2-m_2^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

Special cases:

id	Int	Description

bb0 (dbb0)	1 (19)	(derivative of) scalar two-point one-loop function

bb1 (dbb1)	4 (22)	(derivative of) coefficient of $p_{μ}$

bb00 (dbb00)	7 (25)	(derivative of) coefficient of $g_{μν}$

bb11 (dbb11)	10 (28)	(derivative of) coefficient of $p_μ p_ν$

bb001 (dbb001)	13 (31)	(derivative of) coefficient of $g_{μν}p_ρ$

bb111	16	coefficient of $p_μ p_ν p_ρ$

Functions B0, B1, B00, B11, B001 and B111 are defined.

LoopTools.B11 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.B111 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.Bget — Function

Bget(p^2, m1^2, m2^2; val_only = false)

return the finite piece of all two-point coefficients. See also bget.

val_only = false, return a NamedTuple; otherwise return an NTuple.

LoopTools.C0 — Method

C0(p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2)

the scalar three-point one-loop function

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{d^D q }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

LoopTools.C0i — Method

C0i(id, p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2)

three-point tensor coefficient for id

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{({\rm numerator})\, d^D q }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

Special cases:

id	Int	Description

cc0	1	scalar three-point one-loop function

cc1	4	coefficient of $k_{1μ}$

cc2	7	coefficient of $k_{2μ}$

cc00	10	coefficient of $g_{μν}$

cc11	13	coefficient of $k_{1μ} k_{1ν}$

cc12	16	coefficient of $k_{1μ} k_{2ν}$

cc22	19	coefficient of $k_{2μ} k_{2ν}$

...	...	...

cc2222	64	coefficient of $k_{2μ} k_{2ν} k_{2ρ} k_{2σ}$

where $k_{1,2}$ are related to $p_{1,2}$ by $k_{j} = \sum_{i=1}^j p_i$.

LoopTools.Cget — Function

Cget(p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2; val_only = false)

return the finite piece of all three-point coefficients. See also cget.

val_only = false, return a NamedTuple; otherwise return an NTuple.

LoopTools.D0 — Method

D0(p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, m1^2, m2^2, m3^2, m4^2)

the scalar four-point one-loop function

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{d^D q }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right] \left[(q+p_1+p_2+p_3)^2-m_4^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

LoopTools.D0i — Method

D0i(id, p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, m1^2, m2^2, m3^2, m4^2)

four-point tensor coefficient for id

\[\frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int \frac{({\rm numerator})\, d^D q }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right] \left[(q+p_1+p_2+p_3)^2-m_4^2\right]}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

Special cases:

id	Int	Description

dd0	1	scalar four-point one-loop function

dd1	4	coefficient of $k_{1μ}$

dd2	7	coefficient of $k_{2μ}$

dd3	10	coefficient of $k_{3μ}$

dd00	13	coefficient of $g_{μν}$

...	...	...

dd33333	238	coefficient of $k_{3μ} k_{3ν} k_{3ρ} k_{3σ} k_{3λ}$

where $k_{1,2,3}$ are related to the external momenta $p_{1,2,3}$ by $k_{j} = \sum_{i=1}^j p_i$.

LoopTools.DB0 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.DB00 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.DB1 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.DB11 — Function

Coefficients of two-point tensor loop integral. See B0i.

LoopTools.DRResult — Method

DRResult(c0, c1, c2)

arrange the coefficients of $1/ε$, corresponding to the Mathematica version of the same name. 1/ε corresponds to DR1eps in the Mathematica version.

LoopTools.Dget — Function

Dget(p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, 
m1^2, m2^2, m3^2, m4^2; val_only = false)

return the finite piece of all four-point coefficients. See also dget.

val_only = false, return a NamedTuple; otherwise return an NTuple.

LoopTools.E0 — Method

E0(p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, (p2+p3)^2, (p3+p4)^2, (p4+p5)^2, 
(p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2)

the scalar five-point one-loop function

\[\begin{aligned} \frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int d^D q & \frac{1 }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right] \left[(q+p_1+p_2+p_3)^2-m_4^2\right]} \\ &\times \frac1{\left[(q+p_1+p_2+p_3+p_4)^2-m_5^2\right]} \end{aligned}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

LoopTools.E0i — Method

E0i(id, p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, (p2+p3)^2, (p3+p4)^2, (p4+p5)^2, 
(p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2)

five-point tensor coefficient for id

\[\begin{aligned} \frac{μ^{4-D}}{iπ^{D/2} r_Γ} \int d^D q & \frac{({\rm numerator}) }{(q^2-m_1^2)\left[(q+p_1)^2-m_2^2\right] \left[(q+p_1+p_2)^2-m_3^2\right] \left[(q+p_1+p_2+p_3)^2-m_4^2\right]} \\ &\times \frac1{\left[(q+p_1+p_2+p_3+p_4)^2-m_5^2\right]} \end{aligned}\]

with $r_Γ = \frac{Γ^2(1-ε)Γ(1+ε)}{Γ(1-2ε)}$, $D=4-2ε$.

Special cases:

id	Int	Description

ee0	1	scalar five-point one-loop function

ee1	4	coefficient of $k_{1μ}$

ee2	7	coefficient of $k_{2μ}$

ee3	10	coefficient of $k_{3μ}$

ee00	13	coefficient of $g_{μν}$

...	...	...

ee4444	256	coefficient of $k_{4μ} k_{4ν} k_{4ρ} k_{4σ}$

where $k_{1,2,3,4}$ are related to the external momenta $p_{1,2,3,4}$ by $k_{j} = \sum_{i=1}^j p_i$.

LoopTools.Eget — Function

Eget(p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, (p2+p3)^2, (p3+p4)^2, (p4+p5)^2, 
(p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2; val_only = false)

return the finite piece of all five-point coefficients. See also eget.

val_only = false, return a NamedTuple; otherwise return an NTuple.

LoopTools.Li2 — Method

Li2(x)

return the dilogarithm of $x$.

LoopTools.Li2omx — Method

Li2omx(x)

return the dilogarithm of $1-x$.

LoopTools.aget — Function

aget(msq)

return all one-point coefficients; each one is characterized by three numbers, with the later two coefficients of $ε^{-1}$ and $ε^{-2}$, respectively.

See also Aget and agetsym.

LoopTools.agetsym — Function

agetsym(msq)

return all one-point coefficients with terms of $ε^{-1}$ and $ε^{-2}$ kept symbolically.

LoopTools.aput! — Function

aput!(res::Vector{ComplexF64}, msq)

return all one-point coefficients to the preallocated array res of length 6.

See also aget and agetsym.

LoopTools.bget — Function

bget(p^2, m1^2, m2^2)

return all two-point coefficients; each one is characterized by three numbers, with the later two coefficients of $ε^{-1}$ and $ε^{-2}$, respectively.

See also Bget and bgetsym.

LoopTools.bgetsym — Function

bgetsym(p^2, m1^2, m2^2)

return all two-point coefficients with terms of $ε^{-1}$ and $ε^{-2}$ kept symbolically.

LoopTools.bput! — Function

bput!(res::Vector{ComplexF64}, p^2, m1^2, m2^2)

return all two-point coefficients to the preallocated array res of length 33.

See also bget and bgetsym.

LoopTools.cget — Function

cget(p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2)

return all three-point coefficients; each one is characterized by three numbers, with the later two coefficients of $ε^{-1}$ and $ε^{-2}$, respectively.

See also Cget and cgetsym.

LoopTools.cgetsym — Function

cgetsym(p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2)

return all three-point coefficients with terms of $ε^{-1}$ and $ε^{-2}$ kept symbolically.

LoopTools.clearcache — Function

clearcache()

remove all integrals from the cache.

LoopTools.cput! — Function

cput!(res::Vector{ComplexF64}, p1^2, p2^2, (p1+p2)^2, m1^2, m2^2, m3^2)

return all three-point coefficients to the preallocated array res of length 66.

See also cget and cgetsym.

LoopTools.dget — Function

dget(p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, m1^2, m2^2, m3^2, m4^2)

return all four-point coefficients; each one is characterized by three numbers, with the later two coefficients of $ε^{-1}$ and $ε^{-2}$, respectively. See also Dget and dgetsym.

LoopTools.dgetsym — Function

dgetsym(p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, m1^2, m2^2, m3^2, m4^2)

return all four-point coefficients with terms of $ε^{-1}$ and $ε^{-2}$ kept symbolically.

LoopTools.dput! — Function

dput!(res::Vector{ComplexF64}, p1^2, p2^2, p3^2, p4^2, (p1+p2)^2, (p2+p3)^2, 
m1^2, m2^2, m3^2, m4^2)

return all four-point coefficients to the preallocated array res of length 240.

See also dget and dgetsym.

LoopTools.eget — Function

eget(p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, (p2+p3)^2, (p3+p4)^2, (p4+p5)^2, 
(p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2)

return all five-point coefficients; each one is characterized by three numbers, with the later two coefficients of $ε^{-1}$ and $ε^{-2}$, respectively. See also Eget and egetsym.

LoopTools.egetsym — Function

egetsym(p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, (p2+p3)^2, (p3+p4)^2, (p4+p5)^2, 
(p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2)

return all five-point coefficients with terms of $ε^{-1}$ and $ε^{-2}$ kept symbolically.

LoopTools.eput! — Function

eput!(res::Vector{ComplexF64}, p1^2, p2^2, p3^2, p4^2, p5^2, (p1+p2)^2, 
(p2+p3)^2, (p3+p4)^2, (p4+p5)^2, (p5+p1)^2, m1^2, m2^2, m3^2, m4^2, m5^2)

return all five-point coefficients to the preallocated array res of length 258.

See also eget and egetsym.

LoopTools.getcmpbits — Function

getcmpbits()

the number of bits for the cache-lookup precision. The default is 62 for double precision, 64 for quadruple precision.

LoopTools.getdebugkey — Function

getdebugkey()
setdebugkey()

Debugging output can be turned on.

LoopTools.getdelta — Function

getdelta()
setdelta(Δ::Real)

getdelta() returns the value of $\Delta$, the finite piece of UV divergence, default to 0 (the $\overline{MS}$ scheme); change by setdelta(Δ).

LoopTools.getdiffeps — Function

getdiffeps()
setdiffeps(b::Real)

Two quantities are tested for equality if the absolute value of their difference is smaller than getdiffoeps(). Default value: $10^{-12}$.

LoopTools.getlambda — Function

getlambda()
setlambda(λ::Real)

getdelta() returns the value of $\lambda^2$ for regularizing the IR divergence. It is positive for photon-mass regularization; $0$, $-1$ and $-2$ give the finite piece, coefficients of the $ε^{-1}$ and $ε^{-2}$, respectively. Change by setlambda(λ).

LoopTools.getmaxdev — Function

getmaxdev()
setmaxdev(b::Real)

set the tolerance of the deviation between different implementations of functions. The default value is 1e-12 for issuing warnings.

LoopTools.getminmass — Function

getminmass()
setminmass(m::Real)

For regularizing collinear divergence.

LoopTools.getmudim — Function

getmudim()
setmudim(μ::Real)

The dimensional regularization scale $μ$; the default value is 1.0.

LoopTools.getuvdiv — Function

getuvdiv()
setuvdiv(x::Real)

Switch on (x=1, default) or off (x=0) the UV divergence part of $1/ε$.

LoopTools.getversionkey — Function

getversionkey()

return the version key; the loop integrals are implemented using different versions (Denner, FF). For more information, see Section 1.3.10 of the LoopTools manual.

LoopTools.getzeroeps — Function

getzeroeps()
setzeroeps(b::Real)

A given quantity x is tested to be zero if |x| is smaller than getzeroeps(). Default value: $10^{-22}$.

LoopTools.ltexi — Function

ltexi()

summary of errors.

LoopTools.markcache — Function

markcache()

mark the current cache pointers.

LoopTools.restorecache — Function

restorecache()

restore cache pointers marked by markcache().

LoopTools.setcmpbits — Function

setcmpbits(b::Int)

set the number of bits for cache-lookup precision.

LoopTools.setdebugkey — Function

getdebugkey()
setdebugkey()

Debugging output can be turned on.

LoopTools.setdebugrange — Method

setdebugrange(f::Int, t::Int)

The integrals are listed in the output with a unique serial number. If the list of integrals becomes too long, one can select only a range of serial numbers for viewing.

LoopTools.setdelta — Function

getdelta()
setdelta(Δ::Real)

getdelta() returns the value of $\Delta$, the finite piece of UV divergence, default to 0 (the $\overline{MS}$ scheme); change by setdelta(Δ).

LoopTools.setdiffeps — Function

getdiffeps()
setdiffeps(b::Real)

Two quantities are tested for equality if the absolute value of their difference is smaller than getdiffoeps(). Default value: $10^{-12}$.

LoopTools.setlambda — Function

getlambda()
setlambda(λ::Real)

getdelta() returns the value of $\lambda^2$ for regularizing the IR divergence. It is positive for photon-mass regularization; $0$, $-1$ and $-2$ give the finite piece, coefficients of the $ε^{-1}$ and $ε^{-2}$, respectively. Change by setlambda(λ).

LoopTools.setmaxdev — Function

getmaxdev()
setmaxdev(b::Real)

set the tolerance of the deviation between different implementations of functions. The default value is 1e-12 for issuing warnings.

LoopTools.setminmass — Function

getminmass()
setminmass(m::Real)

For regularizing collinear divergence.

LoopTools.setmudim — Function

getmudim()
setmudim(μ::Real)

The dimensional regularization scale $μ$; the default value is 1.0.

LoopTools.setuvdiv — Function

getuvdiv()
setuvdiv(x::Real)

Switch on (x=1, default) or off (x=0) the UV divergence part of $1/ε$.

LoopTools.setversionkey — Function

setversionkey(k::Int)

set the version key to choose the implementations of various functions and whether checking is performed.

LoopTools.setzeroeps — Function

getzeroeps()
setzeroeps(b::Real)

A given quantity x is tested to be zero if |x| is smaller than getzeroeps(). Default value: $10^{-22}$.